requirements/default.txt
Django==3.0.5
psycopg2==2.8.4
gunicorn==20.0.4
pytz==2019.3
celery==4.4.2
redis==3.4.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
flower==0.9.4
django-debug-toolbar==2.2
requests==2.23.0
djangorestframework==3.11.0
django-rest-swagger==2.2.0
django-silk==4.0.1
